[FEATURE] shop 인증 정보 검증 로직 구현

+++ src/main/java/com/nextroom/oescape/service/HintService.java
@@ -3,17 +3,18 @@ package com.nextroom.oescape.service;
 import static com.nextroom.oescape.exceptions.StatusCode.*;
 
 import java.util.List;
+import java.util.Objects;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.nextroom.oescape.domain.Hint;
-import com.nextroom.oescape.domain.Shop;
 import com.nextroom.oescape.domain.Theme;
 import com.nextroom.oescape.dto.HintDto;
 import com.nextroom.oescape.exceptions.CustomException;
 import com.nextroom.oescape.repository.HintRepository;
 import com.nextroom.oescape.repository.ThemeRepository;
+import com.nextroom.oescape.security.SecurityUtil;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,9 +25,7 @@ public class HintService {
     private final ThemeRepository themeRepository;
 
     @Transactional
-    public void addHint(Shop shop, HintDto.AddHintRequest request) {
-        validateTheme(shop, request.getThemeId());
-
+    public void addHint(HintDto.AddHintRequest request) {
         Theme theme = themeRepository.findById(
                 request.getThemeId()) // TODO optimize by making method get theme from shop
             .orElseThrow(() -> new CustomException(THEME_NOT_FOUND));
@@ -40,50 +39,48 @@ public class HintService {
             .progress(request.getProgress())
             .build();
 
+        if (!Objects.equals(theme.getShop().getAdminCode(), SecurityUtil.getRequestedShopAdminCode())) {
+            throw new CustomException(NOT_PERMITTED);
+        }
+
         hintRepository.save(hint);
     }
 
     @Transactional(readOnly = true)
-    public List<HintDto.HintListResponse> getHintList(Shop shop, Long themeId) {
-        validateTheme(shop, themeId);
-
+    public List<HintDto.HintListResponse> getHintList(Long themeId) {
         Theme theme = themeRepository.findById(
                 themeId) // TODO optimize by making method get theme from shop
             .orElseThrow(() -> new CustomException(THEME_NOT_FOUND));
+
+        if (!Objects.equals(theme.getShop().getAdminCode(), SecurityUtil.getRequestedShopAdminCode())) {
+            throw new CustomException(NOT_PERMITTED);
+        }
 
         return theme.getHints().stream().map(Hint::toHintListResponse).toList();
     }
 
     @Transactional
-    public void editHint(Shop shop, HintDto.EditHintRequest request) {
+    public void editHint(HintDto.EditHintRequest request) {
         Hint hint = hintRepository.findById(request.getId()).orElseThrow(
             () -> new CustomException(HINT_NOT_FOUND)
         );
+
+        if (!Objects.equals(hint.getTheme().getShop().getAdminCode(), SecurityUtil.getRequestedShopAdminCode())) {
+            throw new CustomException(TARGET_HINT_NOT_FOUND);
+        }
+
         hint.update(request);
     }
 
-    public void removeHint(Shop shop, HintDto.RemoveHintRequest request) {
+    public void removeHint(HintDto.RemoveHintRequest request) {
         Hint hint = hintRepository.findById(request.getId()).orElseThrow(
             () -> new CustomException(HINT_NOT_FOUND)
         );
+
+        if (!Objects.equals(hint.getTheme().getShop().getAdminCode(), SecurityUtil.getRequestedShopAdminCode())) {
+            throw new CustomException(TARGET_HINT_NOT_FOUND);
+        }
+
         hintRepository.delete(hint);
-    }
-
-    private void validateTheme(Shop shop, Long themeId) {
-        if (!themeRepository.existsByIdAndShop(themeId, shop)) {
-            throw new CustomException(THEME_NOT_FOUND);
-        }
-    }
-
-    private void validateHint(Shop shop, Long hintId) {
-        Boolean validity = false;
-        for (Theme theme : shop.getThemes()) {
-            if (hintRepository.existsById(hintId)) {
-                validity = true;
-            }
-        }
-        if (!validity) {
-            throw new CustomException(HINT_NOT_FOUND);
-        }
     }
 }
